+++ src/FunFair.BuildVersion/Program.cs
@@ -114,8 +114,10 @@ internal static class Program
         {
             Console.WriteLine($"{VersionInformation.Product} {VersionInformation.Version}");
 
-            return await Parser
-                .Default.ParseArguments<Options>(args)
+            using Parser parser = new(settings => settings.HelpWriter = Console.Out);
+
+            return await parser
+                .ParseArguments<Options>(args)
                 .MapResult(parsedFunc: ParsedOkAsync, notParsedFunc: NotParsedAsync);
         }
         catch (Exception exception)
@@ -175,7 +177,7 @@ internal static class Program
         return new(found);
     }
 
-    private static int FindBuildNumber(int buildNumberFromCommandLine)
+    internal static int FindBuildNumber(int buildNumberFromCommandLine)
     {
         if (buildNumberFromCommandLine > 0)
         {