+++ CHANGELOG.md
@@ -9,9 +9,12 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ## [Unreleased]
 ### Security
 ### Added
+- Added unit tests for FunFair.BuildVersion to increase code coverage to 100%
 ### Fixed
+- Fixed CommandLine Parser not being disposed after argument parsing
 ### Changed
 - SDK - Updated DotNet SDK to 10.0.300
+- Updated FunFair.Test.Common to 6.2.25.2243 in all test projects
 ### Deprecated
 ### Removed
 ### Deployment Changes